moved flask app to subdirectory
