--- .gitignore
@@ -1,8 +1,0 @@
-.idea
-.code
-__pycache__
-photos
-data
-.env
-.venv
-
