+++ data_acquisition/.env-local-base
@@ -6,7 +6,7 @@ SERVER_PORT=8080
 MONGO_URL=localhost:27017
 
 # Change this to any suitable place
-PHOTO_DICT=../photos
+PHOTO_DICT=./photos
 
 SECRET_KEY=123
 API_KEY=456
