+++ data_acquisition/Dockerfile
@@ -10,6 +10,7 @@ ENV SERVER_PORT=8080
 ENV PHOTO_DICT=/photos 
 ENV SECRET_KEY=123
 ENV API_KEY=456
+ENV SESSION_LIFETIME_DAYS=30
 
 WORKDIR /app
