+++ data_acquisition/app/domain/common/authentication/static/js/logout_request.js
