+++ data_acquisition/app/domain/training_data/dto/training_data_dto.py
