+++ data_acquisition/app/domain/training_data/queries/create_command.py
