+++ data_acquisition/app/domain/training_data/static/js/delete_request.js
