+++ data_acquisition/app/domain/training_data/templates/new_datapoint_form.html
