+++ data_acquisition/app/domain/training_data/templates/upload_success.html
