+++ data_acquisition/app/domain/training_data/templates/view_all_data.html
