+++ data_acquisition/app/templates/base.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <title>CAM-Classifier</title>
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="{{ url_for('static', filename='main.css') }}">
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}">
     <script src="{{ url_for('authentication_bp.static', filename='js/logout_request.js') }}"></script>
     <script src="{{ url_for('training_data_bp.static', filename='js/delete_request.js') }}"></script>
 </head>
